Rename exiv2app files to param
