+++ src/actions.hpp
@@ -29,7 +29,7 @@
 
 // *****************************************************************************
 // included header files
-#include "exiv2app.hpp"
+#include "params.hpp"
 
 // *****************************************************************************
 // class declarations
